+++ posce/comms/repl.py
@@ -1,0 +1,20 @@
+'''
+Click command function 'repl'.
+'''
+
+import code
+
+import click
+
+from posce            import VERSION_STRING
+from posce.comms.base import group
+
+@group.command(hidden=True)
+@click.pass_obj
+def repl(book):
+    '''
+    Open a debug REPL.
+    '''
+
+    args = {'book': book, 'group': group}
+    code.interact(banner=VERSION_STRING, local=args, exitmsg='')